[COOK-3608]
The split mail.* facilities in the `50-default.conf` template is a problem
for me on mail servers.  This seems to be an opinionated addition to
Debian-ish rsyslog defaults as shipped, however I retained that
functionality on non-RHEL systems for cookbook backwards compatibility.

Also, the `50-default.conf` is very much geared towards Ubuntu/Debian.
Which is not a huge problem, but the cookbook generates log files which
will never get rotated on RHEL family, and probably other, systems.

I am adding a hash knob to allow log facilities and log destination to
be rendered in the `50-default.conf` template for these two requirements.