Fixes broken tests for suse and adds in license accept
This commit removes opensuse leap 42 as it is no longer supported and
replaces it with opensuse leap 15.

It also adds in opensuseleap to the platform family specific
attributes alongside suse so that the correct attributes are used in the
recipe when the platform_family is opensuseleap.

It also adds in CHEF_LICENSE=accept for the last travis tests.

Signed-off-by: Rudi Kramer <rudi.kramer@fnb.co.za>